tests/wpt/tests/layout-instability/multi-clip-visual-rect.html

<!DOCTYPE html>
<title>Layout Instability: multi clip visual rect</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<style>

body { margin: 0; }
#outer { width: 200px; height: 600px; overflow: hidden; }
#inner { width: 300px; height: 150px; overflow: hidden; }
#j { position: relative; width: 300px; height: 600px; background: blue; }

</style>
<div id='outer'><div id='inner'><div id='j'></div></div></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>

promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  document.querySelector("#j").style.top = "-200px";

  // Note that, while the element moves up 200px, its visibility is
  // clipped at 0px,150px height, so the additional 200px of vertical
  // move distance is not included in the score.
  // (clip width 200) * (clip height 150)
  const expectedScore = computeExpectedScore(200 * 150, 200);

  await watcher.promise;
  assert_equals(watcher.score, expectedScore);
}, "Multi clip visual rect.");

</script>
